modified type of Version from long to int

+++ src/client/gov/epa/emissions/framework/client/data/editor/EditablePageRowSource.java
@@ -23,7 +23,7 @@ public class EditablePageRowSource implements RowSource {
 
         list.addAll(source.tokens());
         list.add(new Integer(source.getRecordId()));
-        list.add(new Long(source.getVersion()));
+        list.add(new Integer(source.getVersion()));
         
 
         return list.toArray();
